backend/requirements.txt
# --- ✅ Core backend framework ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0

# --- ✅ PyTorch + Vision ---
torch==2.5.0
torchvision==0.20.0

# --- ✅ NLP + Google Cloud NLP ---
transformers==4.40.0
google-cloud-language==2.11.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0

# --- ✅ Image Processing ---
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# --- ✅ Scientific Utils ---
numpy==1.26.4
scikit-learn==1.4.2
tqdm==4.66.4

# --- ✅ Optional: DB / Redis (Commented) ---
# sqlalchemy==2.0.23
# alembic==1.13.1
# redis==5.0.1
# aioredis==2.0.1
